:bug: fix: 디렉토리 구조 변경 및 BaseEntity 추가
