+++ src/main/java/colorful/starbucks/StarbucksColorfulApplication.java
@@ -2,8 +2,10 @@ package colorful.starbucks;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class StarbucksColorfulApplication {
 
     public static void main(String[] args) {
