+++ src/main/java/colorful/starbucks/admin/domain/Color.java
@@ -1,0 +1,30 @@
+package colorful.starbucks.admin.domain;
+
+import colorful.starbucks.common.entity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Color extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "color_id")
+    private Long id;
+
+    private String colorName;
+
+    private int extraFee;
+
+    @Builder
+    private Color(Long id, String colorName, int extraFee) {
+        this.id = id;
+        this.colorName = colorName;
+        this.extraFee = extraFee;
+    }
+}
