+++ src/main/java/colorful/starbucks/admin/domain/ProductTopCategory.java
@@ -1,4 +1,4 @@
-package colorful.starbucks.payments.domain;
+package colorful.starbucks.admin.domain;
 
 import colorful.starbucks.common.entity.BaseEntity;
 import jakarta.persistence.*;
@@ -9,20 +9,19 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED )
-public class Payments extends BaseEntity {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ProductTopCategory extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "payments_id")
+    @Column(name = "product_top_category_id")
     private Long id;
 
-    private String paymentType;
+    private String categoryName;
 
     @Builder
-    private Payments(Long id,
-                     String paymentType) {
+    private ProductTopCategory(Long id, String categoryName) {
         this.id = id;
-        this.paymentType = paymentType;
+        this.categoryName = categoryName;
     }
 }
