+++ src/main/java/colorful/starbucks/admin/domain/Size.java
@@ -1,0 +1,30 @@
+package colorful.starbucks.admin.domain;
+
+import colorful.starbucks.common.entity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Size extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "size_id")
+    private Long id;
+
+    private String sizeName;
+
+    private int extraFee;
+
+    @Builder
+    private Size(Long id, String sizeName, int extraFee) {
+        this.id = id;
+        this.sizeName = sizeName;
+        this.extraFee = extraFee;
+    }
+}
